Delegate error reporting to an error handler
Be able to delegate the error handling. Currently two such
handlers exist. One is the old log output and there is a development handler to pop-up a debugger. The package can not be loaded/migrated easily due the lack of instance variable. This would somehow need to be patched as post-load action..