Change the when/how timers are removed
Before we only triggered work and thought the dispatcher
would remove the timer. This is an issue when working with timers, locks and the debugger is coming up.

In that case timers will expire, the dispatcher will queue work but the main timer loop will notice timers have expired and schedule more work. In fact it will kind of busy loop and fill up the dispatcher queue without making progress.

Now expired timers will be removed and scheduled. In case the dispatcher needs to block at least no more work will be scheduled.