Attempt to simplify the scheduler for Pharo7.0
In Pharo7.0 the doShutdown seems to be stuck from time to time. Avoid shutting down the process and just make sure the delay will not make us stuck. If we don't have any timers the normal inserting will wake-up the loop. By just signalling the semaphore we will make sure to move forward.

Co-authored-by: startupscript <info@moiji-mobile.com>